refactor(bridge): extract participant identity
