+++ whatsrust/lib/main.go
@@ -653,31 +653,6 @@ func GetSelfId(client *whatsmeow.Client) string {
 	return StrFromJid(*client.Store.ID)
 }
 
-func participantMatchesSelf(client *whatsmeow.Client, participant types.GroupParticipant) bool {
-	if client == nil || client.Store == nil || client.Store.ID == nil {
-		return false
-	}
-	self := client.Store.ID.ToNonAD()
-	for _, candidate := range []types.JID{participant.JID, participant.PhoneNumber, participant.LID} {
-		if candidate.IsZero() {
-			continue
-		}
-		if jidsMatchSelf(self, candidate) {
-			return true
-		}
-		if candidate.Server == types.HiddenUserServer && self.Server == types.DefaultUserServer {
-			if phone, err := client.Store.LIDs.GetPNForLID(context.Background(), candidate); err == nil && phone.ToNonAD() == self {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func jidsMatchSelf(self, candidate types.JID) bool {
-	return !candidate.IsZero() && candidate.ToNonAD() == self.ToNonAD()
-}
-
 //export C_GetGroupInfo
 func C_GetGroupInfo(cjid C.JID) C.GroupInfoResult {
 	if client == nil || cjid == nil {
